requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.1
joblib==1.3.2